requirements.txt
APScheduler==3.5.3
argon2-cffi==18.3.0
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.10
bcrypt==3.1.5
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Charm-Crypto==0.50
Click==7.0
colorama==0.3.9
coverage==4.5.2
cryptography==2.3.1
defusedxml==0.5.0
Flask==1.0.2
Flask-MQTT==1.0.4
flask-oidc==1.4.0
Flask-OpenID==1.2.5
Flask-Psycopg2==1.3
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httplib2==0.12.0
idna==2.7
isort==4.3.4
ItsDangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
oauth2client==4.1.3
paho-mqtt==1.4.0
passlib==1.7.1
pbr==5.1.0
pg8000==1.12.3
pluggy==0.7.1
psycopg2==2.7.3.2
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pylint==2.1.1
pyope==0.2.2
pyparsing==2.2.1
pytest==3.8.0
pytest-cov==2.6.0
pytest-mock==1.10.0
python3-openid==3.1.0
pytz==2018.6
requests==2.20.0
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
typed-ast==1.1.0
typing==3.6.6
tzlocal==1.5.1
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
